Major refactoring of masking HTML for KnowWE: * Masking now happens with a random masking key, generated uniquely for each new rendering request (instead of a static key). * Renderers now pass a new attribute RenderResult instead of a StringBuilder. It allows to directly mask HTML while appending, instead of masking with a static method. * Methods String.maskHTML() and String.unmaskHTML() no longer exists. Masking now has to happen in the RenderResult. * PageAppendHandler and TagHandler now also use RenderResults in their rendering methods (instead of returning a String). * Fixed a lot of unclean HTML masking.
Attention: This commit is not completely polished. Due to the very high amount of changes and refactoring, there might be some last rendering issues I might have missed. For example it is possible, that HTML is not masked an therefore rendered as plain text in the browser. Please check your own markups to make sure they still work (they should, but I was unable to test them all...).
Additional comments for the new class RenderResult will follow, also a maybe a small developer how-to on rendering an masking in KnowWE.