Fixed object info page for objects containing single quote in in their names.
Note:
(fixed this problem multiple times in KnowWE)
When creating html code that contains single quoted attributes (e.g. <span id='...'>) with single quotes inside the attribute, it is sufficient to use "Stings.encodeHtml(...)" for the content. But if the resulting code is places inside a javascript string and then evaluated later on (or added to the document or used as a javascript-action-link (<a href='javascript:...'>), then the single quote inside the attribute requires to be replaced by "attr.replace("'", "\\'")". Otherwise the html parser decodes the attribute content, but the javascript evaluation does not see the encoded quote any more.