Extended table editor to shift selection by ALT+Cursor (similar to line shift of text fields).
By default, the selection is extended to full lines / rows that are shifted. By additionally pressing Cmd/Ctrl, the selected cells are not extended. The adjacent cells going to be overwritten by moving the selections are "flowing around" these cells and attached at the other side of these cells after shifting.