Fixed FlowchartPropertyTest by adding a second pass for the PropertyDeclarationHandler.
Reason (copied from comment):		
We need a second pass of the handler for named objects, that are not yet created at default priority.
An example for such named objects are Flows. They are created with all completed nodes and edges. The nodes
and edges in turn contain actions which may rely on certain properties again (like units for QuestionNums).
Because Flow ultimately may have to wait for the creation of some properties, they are created after the first
pass. The second pass then will fix those property declarations for named objects like Flows, that could not
be created at the first pass, because they had to wait for this first pass...