fixed rendering of drop-targets
NOTE: was not sure why even texts are checked to be above length of "3". I assumed it was due to non-breaking-spaces used by table editor, whicht are trimmed by Strings.trim(...), so i used that method. So I reworked this, but not finally sure if there are some circumstances I haven't considered.

So please check the modification if any drop-target is missing!