Add button save-as when loading dialog from file
The semantics of the buttons slightly changes when loading from file. The original save button will save to the same file the dialog was loaded from. The new button save-as will do the same as the normal save button when not loading from file (open file chooser).
Also improved styling a bit.
This fixes KNOWWE-29.